pyproject.toml
[project]
name = "vehicle-analyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "transformers (>=4.50.0,<5.0.0)",
    "numpy (>=1.23.0,<2.1.1)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "ultralytics (>=8.3.94,<9.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
